Rename Carousel main entry to index.js
Simplify import paths
