+++ app/components/common/Carousel/carousel.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import { describe, it, expect, vi } from 'vitest';
 import { waitFor } from '@testing-library/react';
 
-import Carousel from './Carousel';
+import Carousel from '.';
 
 const scrollTo = vi.fn();
