+++ app/content/utils/mdx.ts
@@ -17,7 +17,7 @@ import { Section } from 'app/components/common/Section';
 import { VideoCaptionBlock } from 'app/components/common/VideoCaptionBlock';
 import { ImageCaptionBlock } from 'app/components/common/ImageCaptionBlock';
 import { Title } from 'app/components/common/Title';
-import Carousel from 'app/components/common/Carousel/Carousel';
+import Carousel from 'app/components/common/Carousel';
 import dynamic from 'next/dynamic';
 
 const ClientOnly = dynamic(() => import('app/components/common/ClientOnly'), {
